move actions into instrumentation
note that some issues revolving dyngo remain
