+++ instrumentation/appsec/trace/httptrace/http.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2024 Datadog, Inc.
 
-package httpsec
+package httptrace
 
 import (
 	"net/http"
